Update MerlinAU.sh
Code changes & improvements to handle the case where the expected USB-attached drive path set up to download the ZIP file is not found at the time the F/W update check is run. In other situations, if the directory is not located we select the $HOME path and ask the user to confirm when running in "menu mode." Otherwise, we continue wit the temporary selection made.