Update MerlinAU.sh
Added/modified code to output some trace log messages when the "_DoCleanUp_" & "_Reset_LEDs_" functions are called after the F/W Update process is completed. This should leave a trace indicating IF/WHEN the "cleanup" is done as expected, especially if the router reboots itself.

This trace log is only temporary. We can switch it off later on once we have determined whether the "trap" mechanism is working after the F/W Update.