Update MerlinAU.sh
Added a new menu item in "Advanced Options" to give users the choice to select the email format type ("HTML" or "Plain Text"). Default type is HTML which should work with the well-known email servers. If any problems are observed with "garbage lines" getting inserted in the email contents, users can try selecting the "Plain Text" format and test again.