Fixes & improvements with the F/W Update notification date & time.
Fixes & improvements to clarify that a F/W Update occurs on or after the scheduled date & time given the postponement days and the last F/W Notification update. The message & menu now show both the date & time of the F/W Update notification. Currently, the time resolution is in hours, so minutes & seconds are ignored.