Improvements to Acquire Lock
Code improvements when the script is trying to obtain a lock file by allowing a 2nd instance to wait for the 1st instance to perform its job and exit. This means that 2 or more instances could execute in a staggered pattern and not terminate immediately when a lock file is found.