Messaging Improvements
Reworked the error messages when validating the cron schedule to provide users more information of the new restrictions to explain why some parameters fail to pass the check.

Reworded confirmation prompt when enabling automatic script updates. The idea is to provide the user the schedule details of when the checks will be performed so they are more informed when making the choice. Some people may not want to enable if they don't agree with having a daily check.