Code Improvements
More improvements & fine-tuning, primarily in the code that updates the configuration file from the settings done via the WebGUI page. Checks are now made before updating the directory path of the F/W Update ZIP file to make sure it's valid, existing directory.

Also, some preliminary code to support informing the WebGUI of the failed validation of the directory path.