Validator for F/W Update Directory Path
Added a simple validator for the F/W Update directory path. The WebGUI does a simple validation for expected syntax and valid chars. Upon clicking on the "Save" button, the shell script validates that the path actually exists and is writeable, returning success or failure status back to the WebGUI to inform the user.