Improvements for Updating to WebGUI Version
Various code improvements and fine-tuning for updating to newer & latest 1.4.0 WebGUI version.

The goal is to allow for a "smooth transition" from previous 1.3.10 to new 1.4.0 version with WebGUI support.

The latest changes allow users to update to 1.4.0 using the 1.3.10 CLI menu option to update script or via a direct SSH command.