WebGUI Funtionality Improvement
Added functionality to the WebGUI so that users can optionally select to run a "Router Login Test" when saving the password to verify that the login credentials are indeed valid. This is similar to the option available in the CLI menu where users can select to run the same router login test.