Improvements to support APs
Improvements to support APs

Made changes so that we can handle specific functionality or features based on whether the router is in "Router Mode" (i.e. Primary router) or in "AP Mode." This allows us to determine whether the WebGUI should be mounted or whether the AiMesh nodes should be handled only by the Primary router.