Silence Warnings Related to AiMesh/AP Login
For AiMesh nodes, the password is always the same as the primary with no options to change, so we don't need the passwordloginchecker helper for those.

And for an AP, they can be set to any password that doesn't match the primary, so displaying invalid password for those seems incorrect.